fixing issue with nmesh > 1290: using size_t instead of int for ix0, iy0, iz0

+++ src/mesh.c
@@ -32,14 +32,14 @@ int assign_cic(FLOAT* mesh, const int* nmesh, const FLOAT* positions, const FLOA
   for (size_t ii=0; ii<npositions; ii++) {
     const FLOAT weight = weights[ii];
     const FLOAT *pos = &(positions[ii*NDIM]);
-    int ix0 = ((int) pos[0]) % nmesh[0];
-    int iy0 = ((int) pos[1]) % nmesh[1];
-    int iz0 = ((int) pos[2]) % nmesh[2];
-    //int ix0 = (int) pos[0];
-    //int iy0 = (int) pos[1];
-    //int iz0 = (int) pos[2];
+    size_t ix0 = ((int) pos[0]) % nmesh[0];
+    size_t iy0 = ((int) pos[1]) % nmesh[1];
+    size_t iz0 = ((int) pos[2]) % nmesh[2];
+    //size_t ix0 = (int) pos[0];
+    //size_t iy0 = (int) pos[1];
+    //size_t iz0 = (int) pos[2];
     //if (ix0<0 || ix0>=nmesh[0] || iy0<0 || iy0>=nmesh[1] || iz0<0 || iz0>=nmesh[2]) {
-    //  printf("Index out of range: (ix,iy,iz) = (%d,%d,%d) for (%.3f,%.3f,%.3f)\n",ix0,iy0,iz0,pos[0],pos[1],pos[2]);
+    //  printf("Index out of range: (ix,iy,iz) = (%zu,%zu,%zu) for (%.3f,%.3f,%.3f)\n",ix0,iy0,iz0,pos[0],pos[1],pos[2]);
     //  return -1;
     //}
     FLOAT dx = pos[0] - ix0;
@@ -203,14 +203,14 @@ int read_finite_difference_cic(const FLOAT* mesh, const int* nmesh, const FLOAT*
     // see what's going on and to encourage the compiler to optimize
     // and vectorize the code as much as possible.
     const FLOAT *pos = &(positions[ii*NDIM]);
-    int ix0 = ((int) pos[0]) % nmesh[0];
-    int iy0 = ((int) pos[1]) % nmesh[1];
-    int iz0 = ((int) pos[2]) % nmesh[2];
-    //int ix0 = (int) pos[0];
-    //int iy0 = (int) pos[1];
-    //int iz0 = (int) pos[2];
+    size_t ix0 = ((int) pos[0]) % nmesh[0];
+    size_t iy0 = ((int) pos[1]) % nmesh[1];
+    size_t iz0 = ((int) pos[2]) % nmesh[2];
+    //size_t ix0 = (int) pos[0];
+    //size_t iy0 = (int) pos[1];
+    //size_t iz0 = (int) pos[2];
     //if (ix0<0 || ix0>=nmesh[0] || iy0<0 || iy0>=nmesh[1] || iz0<0 || iz0>=nmesh[2]) {
-    //  printf("Index out of range: (ix,iy,iz) = (%d,%d,%d) for (%.3f,%.3f,%.3f)\n",ix0,iy0,iz0,pos[0],pos[1],pos[2]);
+    //  printf("Index out of range: (ix,iy,iz) = (%zu,%zu,%zu) for (%.3f,%.3f,%.3f)\n",ix0,iy0,iz0,pos[0],pos[1],pos[2]);
     //  flag = 1;
     //  continue;
     //}
@@ -283,14 +283,14 @@ int read_cic(const FLOAT* mesh, const int* nmesh, const FLOAT* positions, FLOAT*
   for (size_t ii=0; ii<npositions; ii++) {
     if (flag) continue;
     const FLOAT *pos = &(positions[ii*NDIM]);
-    int ix0 = ((int) pos[0]) % nmesh[0];
-    int iy0 = ((int) pos[1]) % nmesh[1];
-    int iz0 = ((int) pos[2]) % nmesh[2];
-    //int ix0 = (int) pos[0];
-    //int iy0 = (int) pos[1];
-    //int iz0 = (int) pos[2];
+    size_t ix0 = ((int) pos[0]) % nmesh[0];
+    size_t iy0 = ((int) pos[1]) % nmesh[1];
+    size_t iz0 = ((int) pos[2]) % nmesh[2];
+    //size_t ix0 = (int) pos[0];
+    //size_t iy0 = (int) pos[1];
+    //size_t iz0 = (int) pos[2];
     //if (ix0<0 || ix0>=nmesh[0] || iy0<0 || iy0>=nmesh[1] || iz0<0 || iz0>=nmesh[2]) {
-    //  printf("Index out of range: (ix,iy,iz) = (%d,%d,%d) for (%.3f,%.3f,%.3f)\n",ix0,iy0,iz0,pos[0],pos[1],pos[2]);
+    //  printf("Index out of range: (ix,iy,iz) = (%zu,%zu,%zu) for (%.3f,%.3f,%.3f)\n",ix0,iy0,iz0,pos[0],pos[1],pos[2]);
     //  flag = 1;
     //  continue;
     //}
